requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
pillow==10.3.0
tqdm==4.66.4
matplotlib==3.8.4

# CPU PyTorch
torch==2.2.2
torchvision==0.17.2
